Add space to the end of filter conditions in the query
Fixes #6

Previously only the first filter condition in the query was applied. Due
to the order in which filters are constructed, this was usually `keyword`.

Someone making a search with

```
f = Filters(
    keyword = "paypal",
    timespan = "7d",
    theme = "TAX_ECON_PRICE"
)
```

would get the same results as someone using

```
f = Filters(
    keyword = "paypal",
    timespan = "7d"
)
```

which is incorrect.

This is because the API needs a space between each of the filter conditions
in the query.

Add a space at the end of every filter condition when constructing the
query string. I have tested and the API returns the same results if a
space is at the end of a query with only a single condition, ie.

```
f = Filters(
    keyword = "paypal",
    timespan = "7d"
)
```

and

```
f = Filters(
    keyword = "paypal ",
    timespan = "7d"
)
```

have the same results, so it is safe to always add the space in.