Fix multi repeat OR filter
Any query using this filter failed with the error

`TypeError: argument of type 'int' is not iterable`

This is because filters using OR require brackets around the statements
compared with OR. These brackets must only be used when the filter is
using OR - they'll cause an error if they're present for an AND filter.

Add the brackets in for only the OR filters which allows queries to
complete successfully.

Fixes #15