Adjust type hints for `multi_near`
Pylance wasn't happy with the implicit tuple unpacking in the type hint,
so use the `Unpack` helper to tell it what's going on.

I also renamed the variable for once the list has been passed through
`nears` as Pylance was also struggling to understand that it was a new
variable with the same name as the function argument.