Handle empty responses in timeline search
Some timeline searches return no results - particularly those with a
short timespan. The API response when there's no results is simply `{}`
and not the `{"timeline": []}` that this library was expecting.

This meant users got a KeyError whenever they made a search that
returned new results.

I'm not clear if GDELT changed thier API response format since I
originally wrote the handling for the empty case or if this is simply a
long-running bug that's finally been brought to my notice and addressed.

Fixes #58